requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
PyJWT==2.9.0
cryptography==43.0.1
httpx==0.27.2
pytest==8.3.3
pytest-cov==5.0.0
requests==2.32.3
trio==0.26.2
